refactor(proc-macro): remove deprecated proc macro crates
- Remove oso_proc_macro_2 crate (superseded by main implementation)
- Remove oso_proc_macro_logic_2 crate (superseded by main implementation)
- Clean up workspace dependencies in Cargo.toml
- Update Cargo.lock to reflect removed dependencies

These crates were experimental versions that have been replaced
by the stable implementations in oso_proc_macro and oso_proc_macro_logic.
