+++ oso_dev_util/Cargo.toml
@@ -6,5 +6,6 @@ edition = "2024"
 [dependencies]
 anyhow = "*"
 colored = "*"
-oso_proc_macro_2 = { path = "../oso_proc_macro_2" }
+oso_dev_util_helper = { path = "../oso_dev_util_helper" }
+oso_proc_macro = { path = "../oso_proc_macro" }
 toml = { version = "*", features = ["parse"] }
