+++ oso_dev_util/src/decl_manage/crate_.rs
@@ -1,7 +1,6 @@
 use oso_proc_macro_2::FromPathBuf;
 
 use crate::Rslt;
-use crate::cli::Run;
 use crate::decl_manage::package::Package;
 use crate::decl_manage::package::PackageAction;
 use crate::decl_manage::package::PackageInfo;
@@ -10,6 +9,7 @@ use crate::decl_manage::workspace::Workspace;
 use crate::decl_manage::workspace::WorkspaceAction;
 use crate::decl_manage::workspace::WorkspaceInfo;
 use crate::decl_manage::workspace::WorkspaceSurvey;
+use oso_dev_util_helper::cli::Run;
 use std::ffi::OsStr;
 use std::path::Path;
 use std::path::PathBuf;
