+++ oso_dev_util/src/lib.rs
@@ -68,7 +68,6 @@
 
 use anyhow::Result as Rslt;
 
-pub mod cli;
 #[cfg_attr(doc, aquamarine::aquamarine)]
 /// ```mermaid
 /// flowchart TD
