--- oso_proc_macro_2/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "oso_proc_macro_2"
-version = "0.1.0"
-edition = "2024"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-anyhow = "*"
-colored = "*"
-oso_proc_macro_logic_2 = { path = "../oso_proc_macro_logic_2" }
-proc-macro2 = "*"
-quote = "*"
-syn = { version = "*", features = ["full", "extra-traits"] }
