--- oso_proc_macro_logic_2/README.md
@@ -1,178 +1,0 @@
-# oso_proc_macro_logic_2
-
-The latest generation of procedural macro implementation logic for the OSO operating system project, providing enhanced specification processing and code generation capabilities.
-
-## Status
-
-**Almost Complete** - This is the current generation of macro logic with most functionality implemented and several components nearing completion.
-
-## Overview
-
-`oso_proc_macro_logic_2` represents the evolution of the OSO project's macro-driven development approach. Building on the foundation of `oso_proc_macro_logic`, this crate provides improved performance, better error handling, and enhanced specification processing capabilities.
-
-## Key Improvements Over Previous Generation
-
-### Enhanced Performance
-- **Optimized Parsing**: Faster HTML and specification document processing
-- **Improved Caching**: Better caching mechanisms for downloaded specifications
-- **Memory Efficiency**: Reduced memory usage during code generation
-- **Parallel Processing**: Support for concurrent specification processing
-
-### Better Error Handling
-- **Contextual Errors**: More detailed error messages with specification context
-- **Recovery Mechanisms**: Better error recovery and partial processing capabilities
-- **Diagnostic Information**: Enhanced debugging information for macro failures
-- **User-Friendly Messages**: Clearer error messages for common issues
-
-### Advanced Specification Support
-- **Multiple Formats**: Support for additional specification document formats
-- **Version Management**: Better handling of specification version differences
-- **Cross-References**: Improved handling of cross-referenced specifications
-- **Validation**: Enhanced validation of generated code against specifications
-
-## Architecture
-
-### Core Components
-
-#### Specification Processor
-- **Multi-format Parser**: Handles HTML, PDF, and structured document formats
-- **Content Extraction**: Advanced algorithms for extracting structured data
-- **Cross-reference Resolution**: Automatic resolution of specification cross-references
-- **Version Tracking**: Maintains compatibility across specification versions
-
-#### Code Generator
-- **Template Engine**: Flexible template system for code generation
-- **Type System Integration**: Deep integration with Rust's type system
-- **Optimization Passes**: Code optimization during generation
-- **Documentation Generation**: Automatic documentation from specifications
-
-#### Validation Framework
-- **Compile-time Validation**: Ensures generated code compiles correctly
-- **Runtime Testing**: Generates tests that validate against real hardware
-- **Specification Compliance**: Verifies compliance with original specifications
-- **Cross-platform Testing**: Validates across multiple target architectures
-
-## Features
-
-### Specification-Driven Development
-```rust
-// Generate UEFI protocol definitions from latest specification
-generate_uefi_protocols!(version = "2.10", protocols = ["GraphicsOutput", "SimpleFileSystem"]);
-
-// Generate hardware register definitions from vendor specifications
-generate_hardware_registers!(vendor = "ARM", document = "ARMv8-A", version = "latest");
-```
-
-### Advanced Code Generation
-- **Conditional Compilation**: Generate code based on target architecture and features
-- **Optimization Hints**: Include compiler optimization hints in generated code
-- **Documentation Integration**: Generate comprehensive documentation from specifications
-- **Test Generation**: Automatically generate comprehensive test suites
-
-### Enhanced Debugging
-- **Source Mapping**: Maintain mapping from generated code back to specifications
-- **Debug Symbols**: Include debug information in generated code
-- **Tracing Support**: Built-in tracing for macro execution
-- **Interactive Debugging**: Support for interactive macro debugging
-
-## Dependencies
-
-This crate builds upon the OSO ecosystem while introducing minimal external dependencies:
-
-### Internal Dependencies
-- Integration with `oso_dev_util` for shared utilities
-- Compatibility with existing `oso_proc_macro` interfaces
-- Shared error handling with `oso_error`
-
-### External Dependencies
-- Modern parsing libraries for improved performance
-- Enhanced HTTP client for specification downloads
-- Advanced HTML processing capabilities
-- Improved caching and storage systems
-
-## Migration from Previous Generation
-
-### Compatibility
-- **API Compatibility**: Maintains compatibility with existing macro interfaces
-- **Gradual Migration**: Supports gradual migration from `oso_proc_macro_logic`
-- **Feature Parity**: Provides all features of the previous generation
-- **Performance Improvements**: Significant performance improvements for existing functionality
-
-### Migration Path
-1. **Assessment**: Evaluate current macro usage
-2. **Testing**: Comprehensive testing with new implementation
-3. **Gradual Rollout**: Phase-by-phase migration of macro usage
-4. **Validation**: Verify generated code matches previous implementation
-5. **Optimization**: Take advantage of new performance features
-
-## Development Status
-
-### Completed Features
-- ✅ Core specification processing engine
-- ✅ Enhanced HTML parsing and content extraction
-- ✅ Improved code generation framework
-- ✅ Advanced error handling and diagnostics
-- ✅ Performance optimizations and caching
-
-### Almost Complete
-- 🔄 Advanced specification format support
-- 🔄 Cross-reference resolution system
-- 🔄 Enhanced validation framework
-- 🔄 Documentation generation improvements
-- 🔄 Interactive debugging capabilities
-
-### Planned Features
-- 📋 Machine learning-assisted specification parsing
-- 📋 Automated specification update detection
-- 📋 Advanced code optimization passes
-- 📋 Integration with external development tools
-
-## Performance Characteristics
-
-### Benchmarks
-Compared to the previous generation:
-- **Parsing Speed**: 3-5x faster specification parsing
-- **Memory Usage**: 40-60% reduction in peak memory usage
-- **Cache Efficiency**: 80% cache hit rate for repeated builds
-- **Generation Speed**: 2-3x faster code generation
-
-### Scalability
-- **Large Specifications**: Handles specifications with thousands of pages
-- **Concurrent Processing**: Supports parallel processing of multiple specifications
-- **Memory Management**: Efficient memory usage for large-scale code generation
-- **Build Integration**: Optimized for integration with large build systems
-
-## Testing and Validation
-
-### Comprehensive Test Suite
-- **Unit Tests**: Extensive unit test coverage for all components
-- **Integration Tests**: Full integration testing with real specifications
-- **Performance Tests**: Benchmarks and performance regression testing
-- **Compatibility Tests**: Validation against previous generation outputs
-
-### Continuous Validation
-- **Specification Monitoring**: Automatic detection of specification updates
-- **Regression Testing**: Continuous testing against known good outputs
-- **Cross-platform Validation**: Testing across multiple target platforms
-- **Hardware Validation**: Testing generated code against real hardware
-
-## Contributing
-
-When contributing to `oso_proc_macro_logic_2`:
-
-1. **Understand the Architecture**: Familiarize yourself with the enhanced architecture
-2. **Maintain Performance**: Ensure contributions maintain or improve performance
-3. **Test Thoroughly**: Include comprehensive tests for new functionality
-4. **Document Changes**: Update documentation for new features and changes
-5. **Consider Migration**: Ensure changes support migration from previous generation
-
-## License
-
-MIT OR Apache-2.0
-
-## Related Crates
-
-- `oso_proc_macro_logic`: Previous generation implementation (legacy)
-- `oso_proc_macro`: Public procedural macro interfaces
-- `oso_dev_util`: Shared development utilities
-- `oso_error`: Error handling primitives
