--- oso_proc_macro_logic_2/src/cli.rs
@@ -1,1 +1,0 @@
-../../oso_dev_util/src/cli.rs
